Update dependencies and make code improvements
Updated various dependency versions in the pom.xml file for better compatibility and performance. Utilized more efficient methods in the ExpressionUtils.java and StringUtils.java classes such as using isEmpty() instead of checking if the length is 0 and methods like getFirst() which is more intuitive. Also, handled version details better in application.yml and adjusted the version fetching mechanism in version_pom.